test: create adjust down risk tests for v2
