+++ packages/oasis-actions/src/helpers/calculations/Position.ts
@@ -553,12 +553,12 @@ export class Position implements IPosition {
           `Is flashloan required: ${isFlashloanRequired}`,
           `Amount to flashloan: ${amountToFlashloan}`,
           `----`,
-          `Normalised swap or Swapped Amount: ${normalisedSwapOrSwappedAmount.toString()}`,
+          `Normalised unknown X: ${normalisedSwapOrSwappedAmount.toString()}`,
           `Normalised from token amount: ${normalisedFromTokenAmount.toString()}`,
           `Normalised from token amount after fees: ${normalisedFromTokenAmount
             .minus(normalisedSourceFee)
             .toString()}`,
-          `Swap or Swapped Amount: ${swapOrSwappedAmount.toString()}`,
+          `Unknown X: ${swapOrSwappedAmount.toString()}`,
           `From token amount: ${fromTokenAmount.toString()}`,
           `From token amount after fees: ${fromTokenAmountAfterFee.toString()}`,
           `From token: ${
