+++ packages/oasis-actions/src/strategies/aave/adjust/index.ts
@@ -1,2 +1,1 @@
-export { adjustNew as adjust } from './adjustNew'
-// export { adjust } from './adjust'
+export { adjust } from './adjust'
